server/requirements.txt
pymongo[srv]==4.10.1
dnspython==2.7.0
Flask==3.0.3
python-dotenv==1.0.1
openai==1.61.0
sentence-transformers==5.2.3
transformers==5.2.0
tokenizers==0.22.2
flask-cors==6.0.1
Werkzeug==3.1.3
requests==2.32.5
beautifulsoup4==4.14.3
pandas==3.0.0
numpy==2.4.2
PyJWT==2.8.1
pytest==8.3.5
pytest-asyncio==0.25.0
certifi==2024.2.2